requirements.txt
amqp==5.1.1
APScheduler==3.9.1.post1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autobahn==22.12.1
Automat==22.10.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
black==22.12.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
cfgv==3.3.1
channels==3.0.5
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.0
daphne==3.0.2
distlib==0.3.6
Django==4.1.5
django-apscheduler==0.6.2
django-celery-results==2.4.0
django-compat==1.0.15
django-cors-headers==3.13.0
django-debug-toolbar==3.8.1
django-rest==0.8.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
filelock==3.9.0
flake8==6.0.0
gunicorn==20.1.0
hyperlink==21.0.0
identify==2.5.12
idna==3.4
incremental==22.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
mccabe==0.7.0
mypy-extensions==0.4.3
names==0.3.0
nodeenv==1.7.0
packaging==22.0
pathspec==0.10.3
platformdirs==2.6.2
pre-commit==2.21.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.6.0
pyOpenSSL==23.0.0
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
redis==4.4.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
Twisted==22.10.0
txaio==22.2.1
typing-extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.13
vine==5.0.0
virtualenv==20.17.1
wcwidth==0.2.5
websockets==10.4
zope.interface==5.5.2
